change SamplePacket dataLen to 16bit

+++ include/samplePacket.hpp
@@ -7,17 +7,17 @@
 class SamplePacket {
 private:
 	void *data;
-	uint32_t dataLen;
+	uint16_t dataLen;
 
 public:
 	SamplePacket(const SamplePacket &p) : data(p.data), dataLen(p.dataLen){};
 	SamplePacket() : data(nullptr), dataLen(0){};
-	SamplePacket(void *data, uint32_t dataLen) : data(data), dataLen(dataLen){};
+	SamplePacket(void *data, uint16_t dataLen) : data(data), dataLen(dataLen){};
 	~SamplePacket() { free(data); }
 
 	void *getData() { return data; };
-	uint32_t getDataLen() { return dataLen; };
-	void setDataLen(uint32_t l) { dataLen = l; };
+	uint16_t getDataLen() { return dataLen; };
+	void setDataLen(uint16_t l) { dataLen = l; };
 };
 
 #endif /* SAMPLE_PACKET_HPP */
